Add validations to product and cart routes, change of logic in managers, create validation util functions and restructure directories

+++ src/routes/products.router.js
@@ -1,7 +1,8 @@
 import { Router } from "express"
 import ProductManager from '../ProductManager.js'
+import {productBodyValidationPost, productBodyValidationPut} from '../util.js'
 
-let filePath = './src/'
+let filePath = './src/JSON/'
 let fileName = 'products.json'
 const pm = new ProductManager(filePath+fileName)
 
@@ -10,29 +11,35 @@ const router = Router()
 router.get('/', async (req, res)=>{
     try {
         const products = await pm.getProducts()
-        res.send(products)
+        res.status(200).send({status:"SUCCESS", response: products})
     } catch (error) {
-        res.send(error)
+        res.status(500).send({status:"ERROR", error})
     }
 })
 
 router.get('/:pId', async (req, res)=>{
     try {
         let productId = parseInt(req.params.pId)
+        if(isNaN(productId)){
+            res.status(400).send({status:"ERROR", response: "WARNING: Provide a valid product ID"})
+            return
+        }
         const product = await pm.getProductById(productId)
-        res.send(product)
+        res.status(200).send({status:"SUCCESS", response: product})
     } catch (error) {
-        res.send(error)
+        res.status(500).send({status:"ERROR", error})
     }
 })
 
 router.post('/', async (req, res)=>{
     try {
         let productBody = req.body
-        const response = await pm.addProduct(productBody)
-        res.send(response)
+        const {title, description, code, price, status, stock, category, thumbnails} = productBodyValidationPost(productBody)
+        console.log(title, description, code, price, status, stock, category, thumbnails)
+        const response = await pm.addProduct(title,description, code, price, status, stock, category, thumbnails)
+        res.status(200).send({status:"SUCCESS", response: response})
     } catch (error) {
-        res.send(error)
+        res.status(500).send({status:"ERROR", error})
     }
 })
 
@@ -40,20 +47,25 @@ router.put('/:pId', async (req, res)=>{
     try {
         let productId = parseInt(req.params.pId)
         let productBody = req.body
-        const response = await pm.updateProduct(productId, productBody)
-        res.send(response)
+        let validatedBody = productBodyValidationPut(productBody)
+        const response = await pm.updateProduct(productId, validatedBody)
+        res.status(200).send({status:"SUCCESS", response: response})
     } catch (error) {
-        res.send(error)
+        res.status(500).send({status:"ERROR", error})
     }
 })
 
 router.delete('/:pId', async (req, res)=>{
     try {
         let productId = parseInt(req.params.pId)
+        if(isNaN(productId)){
+            res.status(400).send({status:"ERROR", response: "WARNING: Provide a valid product ID"})
+            return
+        }
         const response = await pm.deleteProduct(productId)
-        res.send(response)
+        res.status(200).send({status:"SUCCESS", response: response})
     } catch (error) {
-        res.send(error)
+        res.status(500).send({status:"ERROR", error})
     }
 })
